requirements.txt
absl-py==1.4.0
attrs==23.1.0
cffi==1.15.1
contourpy==1.0.7
cycler==0.11.0
flatbuffers==23.3.3
fonttools==4.39.3
kiwisolver==1.4.4
matplotlib==3.7.1
mediapipe==0.9.3.0
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
protobuf==3.20.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sounddevice==0.4.6
tenacity==8.2.2
tzdata==2023.3
